# docs/openapi.yaml
openapi: 3.1.0
info:
  title: International Calling API
  version: 0.1.0
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /calls/outbound:
    post:
      tags: [Calls]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutboundCallRequest'
      responses:
        '200':
          description: Call authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboundCallResponse'
  /calls/history:
    get:
      tags: [Calls]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Call history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CallHistoryResponse'
  /webhooks/stripe:
    post:
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Acknowledge webhook
  /webhooks/voice:
    post:
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Acknowledge webhook
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OutboundCallRequest:
      type: object
      properties:
        to: { type: string }
      required: [to]
    OutboundCallResponse:
      type: object
      properties:
        callSessionId: { type: string }
        dialString: { type: string }
        reservedCents: { type: integer }
        currency: { type: string }
        ratePerMinuteCents: { type: integer }
    CallHistoryResponse:
      type: object
      properties:
        id: { type: string }
        direction: { type: string }
        from: { type: string }
        to: { type: string }
        startTs: { type: string, format: date-time }
        endTs: { type: string, format: date-time }
        costCents: { type: integer }
        status: { type: string }
...